changed frames to seconds

+++ wav.h
@@ -10,5 +10,5 @@ You should have received a copy of the CC0 Public Domain Dedication along with t
 #define WAV_H
 #include "pcm.h"
 int open_wav_read(pcm_t **, char *);
-int open_wav_write(pcm_t **, char *, int, int, int);
+int open_wav_write(pcm_t **, char *, int, int, float);
 #endif
